refactor: [#102] extract component for duplicate code
The component for the torrent details was ducplicated for the two URLs:

- http://localhost:3000/torrent/1fc73eb95330bb80251f3fd3b96d1dfd120e81ec/test-test-123-test
- http://localhost:3000/torrent/1fc73eb95330bb80251f3fd3b96d1dfd120e81ec

Depending on whether you use the "slug" with the title or not.
